+++ noodles-vcf/src/writer/builder.rs
@@ -7,50 +7,76 @@ use std::{
 use noodles_bgzf as bgzf;
 
 use super::Writer;
+use crate::io::CompressionMethod;
 
 /// A BAM writer builder.
 #[derive(Debug, Default)]
-pub struct Builder;
+pub struct Builder {
+    compression_method: Option<CompressionMethod>,
+}
 
 impl Builder {
+    /// Sets the compression method.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use noodles_vcf::{io::CompressionMethod, writer::Builder};
+    /// let builder = Builder::default().set_compression_method(CompressionMethod::Bgzf);
+    /// ```
+    pub fn set_compression_method(mut self, compression_method: CompressionMethod) -> Self {
+        self.compression_method = Some(compression_method);
+        self
+    }
+
     /// Builds a VCF writer from a path.
+    ///
+    /// If the compression method is not set, it is detected from the path extension.
     ///
     /// # Examples
     ///
     /// ```no_run
     /// use noodles_vcf as vcf;
-    /// let writer = vcf::writer::Builder.build_from_path("out.vcf")?;
+    /// let writer = vcf::writer::Builder::default().build_from_path("out.vcf")?;
     /// # Ok::<_, std::io::Error>(())
     /// ```
-    pub fn build_from_path<P>(self, dst: P) -> io::Result<Writer<Box<dyn Write>>>
+    pub fn build_from_path<P>(mut self, dst: P) -> io::Result<Writer<Box<dyn Write>>>
     where
         P: AsRef<Path>,
     {
         let dst = dst.as_ref();
 
+        if self.compression_method.is_none() {
+            self.compression_method = match dst.extension().and_then(|ext| ext.to_str()) {
+                Some("gz" | "bgz") => Some(CompressionMethod::Bgzf),
+                _ => Some(CompressionMethod::None),
+            };
+        }
+
         let file = File::create(dst)?;
-
-        let writer: Box<dyn Write> = match dst.extension().and_then(|ext| ext.to_str()) {
-            Some("gz" | "bgz") => Box::new(bgzf::Writer::new(file)),
-            _ => Box::new(BufWriter::new(file)),
-        };
-
-        Ok(Writer::new(writer))
+        Ok(self.build_from_writer(file))
     }
 
     /// Builds a VCF writer from a writer.
+    ///
+    /// If the compression method is not set, no compression is used.
     ///
     /// # Examples
     ///
     /// ```
     /// # use std::io;
     /// use noodles_vcf as vcf;
-    /// let writer = vcf::writer::Builder.build_from_writer(io::sink());
+    /// let writer = vcf::writer::Builder::default().build_from_writer(io::sink());
     /// ```
-    pub fn build_from_writer<W>(self, writer: W) -> Writer<W>
+    pub fn build_from_writer<'w, W>(self, writer: W) -> Writer<Box<dyn Write + 'w>>
     where
-        W: Write,
+        W: Write + 'w,
     {
-        Writer::new(writer)
+        let inner: Box<dyn Write> = match self.compression_method {
+            Some(CompressionMethod::Bgzf) => Box::new(bgzf::Writer::new(writer)),
+            Some(CompressionMethod::None) | None => Box::new(BufWriter::new(writer)),
+        };
+
+        Writer::new(inner)
     }
 }